Interim update, integrating orlando reader
This update adds in an ovOrlandoReader class that uses the vtkAlgorithm
paradigm to read Orlando XML files and output a vtkGraph from the data.
The use of vtkGraph has not yet been incorperated into the QT user
interface so this update (and likely many to come) is not "production"
quality.  In other words, if you're not a developer, don't try using
this particular version of the code.